+++ framework/Furion/TaskScheduler/Attributes/SpareTimeAttribute.cs
@@ -1,5 +1,6 @@
 ﻿using Furion.DependencyInjection;
 using System;
+using System.Text.RegularExpressions;
 
 namespace Furion.TaskScheduler
 {
@@ -9,6 +10,11 @@ namespace Furion.TaskScheduler
     [SkipScan, AttributeUsage(AttributeTargets.Method, AllowMultiple = true)]
     public class SpareTimeAttribute : Attribute
     {
+        /// <summary>
+        /// 模板正则表达式
+        /// </summary>
+        private const string templatePattern = @"\{(?<p>.+?)\}";
+
         /// <summary>
         /// 构造函数
         /// </summary>
@@ -24,12 +30,26 @@ namespace Furion.TaskScheduler
         /// <summary>
         /// 构造函数
         /// </summary>
-        /// <param name="cronExpression"></param>
+        /// <param name="expressionOrKey">表达式或配置Key</param>
         /// <param name="workName"></param>
-        public SpareTimeAttribute(string cronExpression, string workName = default)
+        public SpareTimeAttribute(string expressionOrKey, string workName = default)
         {
-            CronExpression = cronExpression;
-            Type = SpareTimeTypes.Cron;
+            // 支持从配置文件读取
+            var realValue = ResolveTemplate(expressionOrKey);
+
+            // 如果能够转换成整型，则采用间隔
+            if (long.TryParse(realValue, out var interval))
+            {
+                Interval = interval;
+                Type = SpareTimeTypes.Interval;
+            }
+            // 否则当作 Cron 表达式
+            else
+            {
+                CronExpression = realValue;
+                Type = SpareTimeTypes.Cron;
+            }
+
             WorkerName = workName;
         }
 
@@ -77,5 +97,23 @@ namespace Furion.TaskScheduler
         /// Cron 表达式格式化方式
         /// </summary>
         public CronFormat CronFormat { get; set; } = CronFormat.Standard;
+
+        /// <summary>
+        /// 解析模板（可配置 appsetting.json 的键）
+        /// </summary>
+        /// <param name="template"></param>
+        /// <returns></returns>
+        private static string ResolveTemplate(string template)
+        {
+            if (string.IsNullOrWhiteSpace(template)) return default;
+
+            if (Regex.IsMatch(template, templatePattern))
+            {
+                var configKey = Regex.Match(template, templatePattern).Groups["p"].Value;
+                return App.GetConfig<string>(configKey) ?? configKey;
+            }
+
+            return template;
+        }
     }
 }