+++ handbook/src/pages/index.js
@@ -156,7 +156,7 @@ function Gitee() {
             className={"furion-log-jiao" + (isDarkTheme ? " dark" : "")}
           ></div>
           <div className="furion-log-number">
-            <div style={{ color: "#1fd898" }}>164,050</div>
+            <div style={{ color: "#1fd898" }}>165,222</div>
             <span className={isDarkTheme ? " dark" : ""}>Downloads</span>
           </div>
         </div>